gs alias to test only
